iroh_docs::api

Struct Doc

pub struct Doc { /* private fields */ }

Document handle

Implementations

impl Doc

pub fn id(&self) -> NamespaceId

Returns the document id of this doc.

pub async fn close(&self) -> Result<()>

Closes the document.

pub async fn set_bytes( &self, author_id: AuthorId, key: impl Into<Bytes>, value: impl Into<Bytes>, ) -> Result<Hash>

Sets the content of a key to a byte array.

pub async fn set_hash( &self, author_id: AuthorId, key: impl Into<Bytes>, hash: Hash, size: u64, ) -> Result<()>

Sets an entry on the doc via its key, hash, and size.

pub async fn del( &self, author_id: AuthorId, prefix: impl Into<Bytes>, ) -> Result<usize>

Deletes entries that match the given author and key prefix.

This inserts an empty entry with the key set to prefix, effectively clearing all other entries whose key starts with or is equal to the given prefix.

Returns the number of entries deleted.

pub async fn get_exact( &self, author: AuthorId, key: impl AsRef<[u8]>, include_empty: bool, ) -> Result<Option<Entry>>

Returns an entry for a key and author.

Optionally also returns the entry unless it is empty (i.e. a deletion marker).

pub async fn get_many( &self, query: impl Into<Query>, ) -> Result<impl Stream<Item = Result<Entry>>>

Returns all entries matching the query.

pub async fn get_one(&self, query: impl Into<Query>) -> Result<Option<Entry>>

Returns a single entry.

pub async fn share( &self, mode: ShareMode, addr_options: AddrInfoOptions, ) -> Result<DocTicket>

Shares this document with peers over a ticket.

pub async fn start_sync(&self, peers: Vec<NodeAddr>) -> Result<()>

Starts to sync this document with a list of peers.

pub async fn leave(&self) -> Result<()>

Stops the live sync for this document.

pub async fn subscribe( &self, ) -> Result<impl Stream<Item = Result<LiveEvent>> + Send + Unpin + 'static>

Subscribes to events for this document.

pub async fn status(&self) -> Result<OpenState>

Returns status info for this document

pub async fn set_download_policy(&self, policy: DownloadPolicy) -> Result<()>

Sets the download policy for this document

pub async fn get_download_policy(&self) -> Result<DownloadPolicy>

Returns the download policy for this document

pub async fn get_sync_peers(&self) -> Result<Option<Vec<PeerIdBytes>>>

Returns sync peers for this document

pub async fn import_file( &self, blobs: &Store, author: AuthorId, key: Bytes, path: impl AsRef<Path>, import_mode: ImportMode, ) -> Result<ImportFileProgress>

Adds an entry from an absolute file path

pub async fn export_file( &self, blobs: &Store, entry: Entry, path: impl AsRef<Path>, mode: ExportMode, ) -> Result<ExportProgress>

Exports an entry as a file to a given absolute path.