iroh_docs

Module engine

Handlers and actors to for live syncing replicas.

crate::Replica is also called documents here.

Structs

DefaultAuthor

Persistent default author for a docs engine.

Engine

The sync engine coordinates actors that manage open documents, set-reconciliation syncs with peers and a gossip swarm for each syncing document.

ProtectCallbackHandler

The handler for a blobs protection callback.

SyncEvent

Event emitted when a sync operation completes

Enums

DefaultAuthorStorage

Where to persist the default author.

LiveEvent

Events informing about actions of the live sync progress.

Origin

Why we performed a sync exchange

SyncReason

Why we started a sync request