Module iroh_docs::engine::live

Structs

LiveActor

QueuedHashes 🔒

Subscribers 🔒

SubscribersMap 🔒

SyncDetails

SyncEvent

Event emitted when a sync operation completes

SyncReport

Report of a successful sync with the new heads.

Enums

Event

Events informing about actions of the live sync progress.

Op

An iroh-docs operation

ToLiveActor

Messages to the sync actor

Constants

SOURCE_NAME 🔒

Name used for logging when new node addresses are added from the docs engine.

Functions

fmt_accept_namespace 🔒

fmt_accept_peer 🔒

Type Aliases

DownloadRes 🔒

SyncAcceptRes 🔒

SyncConnectRes 🔒