Struct iroh_docs::engine::live::LiveActor

pub struct LiveActor<B: Store> {

    inbox: Receiver<ToLiveActor>,
    sync: SyncHandle,
    endpoint: Endpoint,
    gossip: Gossip,
    bao_store: B,
    downloader: Downloader,
    replica_events_tx: Sender<Event>,
    replica_events_rx: Receiver<Event>,
    sync_actor_tx: Sender<ToLiveActor>,
    gossip_actor_tx: Sender<ToGossipActor>,
    running_sync_connect: JoinSet<(NamespaceId, PublicKey, SyncReason, Result<SyncFinished, ConnectError>)>,
    running_sync_accept: JoinSet<Result<SyncFinished, AcceptError>>,
    download_tasks: JoinSet<(NamespaceId, Hash, Result<Stats, DownloadError>)>,
    missing_hashes: HashSet<Hash>,
    queued_hashes: QueuedHashes,
    subscribers: SubscribersMap,
    state: NamespaceStates,

}

Fields
inbox: Receiver<ToLiveActor>

Receiver for actor messages.

sync: SyncHandleendpoint: Endpointgossip: Gossipbao_store: Bdownloader: Downloaderreplica_events_tx: Sender<Event>replica_events_rx: Receiver<Event>sync_actor_tx: Sender<ToLiveActor>

Send messages to self. Note: Must not be used in methods called from Self::run directly to prevent deadlocks. Only clone into newly spawned tasks.

gossip_actor_tx: Sender<ToGossipActor>running_sync_connect: JoinSet<(NamespaceId, PublicKey, SyncReason, Result<SyncFinished, ConnectError>)>

Running sync futures (from connect).

running_sync_accept: JoinSet<Result<SyncFinished, AcceptError>>

Running sync futures (from accept).

download_tasks: JoinSet<(NamespaceId, Hash, Result<Stats, DownloadError>)>

Running download futures.

missing_hashes: HashSet<Hash>

Content hashes which are wanted but not yet queued because no provider was found.

queued_hashes: QueuedHashes

Content hashes queued in downloader.

subscribers: SubscribersMap

Subscribers to actor events

state: NamespaceStates

Sync state per replica and peer

Implementations

impl<B: Store> LiveActor<B>

pub fn new( sync: SyncHandle, endpoint: Endpoint, gossip: Gossip, bao_store: B, downloader: Downloader, inbox: Receiver<ToLiveActor>, sync_actor_tx: Sender<ToLiveActor>, gossip_actor_tx: Sender<ToGossipActor> ) -> Self

Create the live actor.

pub async fn run(self, gossip_actor: GossipActor) -> Result<()>

Run the actor loop.

async fn run_inner(&mut self) -> Result<Sender<()>>

async fn on_actor_message(&mut self, msg: ToLiveActor) -> Result<bool>

fn sync_with_peer( &mut self, namespace: NamespaceId, peer: PublicKey, reason: SyncReason )

async fn shutdown(&mut self) -> Result<()>

async fn start_sync( &mut self, namespace: NamespaceId, peers: Vec<NodeAddr> ) -> Result<()>

async fn leave( &mut self, namespace: NamespaceId, kill_subscribers: bool ) -> Result<()>

async fn join_peers( &mut self, namespace: NamespaceId, peers: Vec<NodeAddr> ) -> Result<()>

async fn on_sync_via_connect_finished( &mut self, namespace: NamespaceId, peer: PublicKey, reason: SyncReason, result: Result<SyncFinished, ConnectError> )

async fn on_sync_via_accept_finished( &mut self, res: Result<SyncFinished, AcceptError> )

async fn on_sync_finished( &mut self, namespace: NamespaceId, peer: PublicKey, origin: Origin, result: Result<SyncFinished> )

async fn broadcast_neighbors(&self, namespace: NamespaceId, op: &Op)

async fn on_download_ready( &mut self, namespace: NamespaceId, hash: Hash, res: Result<Stats, DownloadError> )

async fn on_neighbor_content_ready( &mut self, namespace: NamespaceId, node: NodeId, hash: Hash )

async fn on_sync_report(&mut self, from: PublicKey, report: SyncReport)

async fn on_replica_event(&mut self, event: Event) -> Result<()>

async fn start_download( &mut self, namespace: NamespaceId, hash: Hash, node: PublicKey, only_if_missing: bool )

pub async fn handle_connection(&mut self, conn: Connecting)

pub fn accept_sync_request( &mut self, namespace: NamespaceId, peer: PublicKey ) -> AcceptOutcome