iroh_docs
Module net

Available on crate feature net only.

Network implementation of the iroh-docs protocol

Structs

SyncFinished

Details of a finished sync operation.

Timings

Time a sync operation took

Enums

AbortReason

Reason why we aborted an incoming sync request.

AcceptError

Errors that may occur on handling incoming sync connections.

AcceptOutcome

Whether we want to accept or reject an incoming sync request.

ConnectError

Errors that may occur on outgoing sync requests.

Constants

ALPN

The ALPN identifier for the iroh-docs protocol

Functions

connect_and_sync

Connect to a peer and sync a replica

handle_connection

Handle an iroh-docs connection and sync all shared documents in the replica store.