Module iroh_docs::ranger

Implementation of Set Reconcilliation based on “Range-Based Set Reconciliation” by Aljoscha Meyer.

Structs

Fingerprint

Message

Range

Stores a range.

RangeFingerprint

RangeItem

Transfers items inside a range to the other participant.

SyncConfig

Enums

InsertOutcome

The outcome of a Store::put operation.

MessagePart

Traits

RangeEntry

Store entries that can be fingerprinted and put into ranges.

RangeKey

A trait constraining types that are valid entry keys.

RangeValue

A trait constraining types that are valid entry values.

Store