Trait iroh_docs::ranger::RangeEntry

pub trait RangeEntry: Debug + Clone {
    type Key: RangeKey;
    type Value: RangeValue;

    // Required methods
    fn key(&self) -> &Self::Key;
    fn value(&self) -> &Self::Value;
    fn as_fingerprint(&self) -> Fingerprint;
}

Store entries that can be fingerprinted and put into ranges.

Required Associated Types

type Key: RangeKey

The key type for this entry.

This type must implement Ord to define the range ordering used in the set reconciliation algorithm.

See RangeKey for details.

type Value: RangeValue

The value type for this entry. See

The type must implement Ord to define the time ordering of entries used in the prefix deletion algorithm.

See RangeValue for details.

Required Methods

fn key(&self) -> &Self::Key

Get the key for this entry.

fn value(&self) -> &Self::Value

Get the value for this entry.

fn as_fingerprint(&self) -> Fingerprint

Get the fingerprint for this entry.

Object Safety

This trait is not object safe.