Trait iroh_docs::ranger::Store

pub trait Store<E: RangeEntry>: Sized {
    type Error: Debug + Send + Sync + Into<Error> + 'static;
    type RangeIterator<'a>: Iterator<Item = Result<E, Self::Error>>
       where Self: 'a,
             E: 'a;
    type ParentIterator<'a>: Iterator<Item = Result<E, Self::Error>>
       where Self: 'a,
             E: 'a;

    // Required methods
    fn get_first(&mut self) -> Result<E::Key, Self::Error>;
    fn get(&mut self, key: &E::Key) -> Result<Option<E>, Self::Error>;
    fn len(&mut self) -> Result<usize, Self::Error>;
    fn is_empty(&mut self) -> Result<bool, Self::Error>;
    fn get_fingerprint(
        &mut self,
        range: &Range<E::Key>
    ) -> Result<Fingerprint, Self::Error>;
    fn entry_put(&mut self, entry: E) -> Result<(), Self::Error>;
    fn get_range(
        &mut self,
        range: Range<E::Key>
    ) -> Result<Self::RangeIterator<'_>, Self::Error>;
    fn prefixed_by(
        &mut self,
        prefix: &E::Key
    ) -> Result<Self::RangeIterator<'_>, Self::Error>;
    fn prefixes_of(
        &mut self,
        key: &E::Key
    ) -> Result<Self::ParentIterator<'_>, Self::Error>;
    fn all(&mut self) -> Result<Self::RangeIterator<'_>, Self::Error>;
    fn entry_remove(&mut self, key: &E::Key) -> Result<Option<E>, Self::Error>;
    fn remove_prefix_filtered(
        &mut self,
        prefix: &E::Key,
        predicate: impl Fn(&E::Value) -> bool
    ) -> Result<usize, Self::Error>;

    // Provided methods
    fn get_range_len(
        &mut self,
        range: Range<E::Key>
    ) -> Result<usize, Self::Error> { ... }
    fn initial_message(&mut self) -> Result<Message<E>, Self::Error> { ... }
    fn process_message<F, F2, F3>(
        &mut self,
        config: &SyncConfig,
        message: Message<E>,
        validate_cb: F,
        on_insert_cb: F2,
        content_status_cb: F3
    ) -> Result<Option<Message<E>>, Self::Error>
       where F: Fn(&Self, &E, ContentStatus) -> bool,
             F2: FnMut(&Self, E, ContentStatus),
             F3: Fn(&Self, &E) -> ContentStatus { ... }
    fn put(&mut self, entry: E) -> Result<InsertOutcome, Self::Error> { ... }

}

Required Associated Types

type Error: Debug + Send + Sync + Into<Error> + 'static

type RangeIterator<'a>: Iterator<Item = Result<E, Self::Error>> where Self: 'a, E: 'a

type ParentIterator<'a>: Iterator<Item = Result<E, Self::Error>> where Self: 'a, E: 'a

Required Methods

fn get_first(&mut self) -> Result<E::Key, Self::Error>

Get a the first key (or the default if none is available).

fn get(&mut self, key: &E::Key) -> Result<Option<E>, Self::Error>

Get a single entry.

fn len(&mut self) -> Result<usize, Self::Error>

Get the number of entries in the store.

fn is_empty(&mut self) -> Result<bool, Self::Error>

Returns true if the vector contains no elements.

fn get_fingerprint( &mut self, range: &Range<E::Key> ) -> Result<Fingerprint, Self::Error>

Calculate the fingerprint of the given range.

fn entry_put(&mut self, entry: E) -> Result<(), Self::Error>

Insert just the given key value pair.

This will replace just the existing entry, but will not perform prefix deletion.

fn get_range( &mut self, range: Range<E::Key> ) -> Result<Self::RangeIterator<'_>, Self::Error>

Returns all entries in the given range.

fn prefixed_by( &mut self, prefix: &E::Key ) -> Result<Self::RangeIterator<'_>, Self::Error>

Returns all entries whose key starts with the given prefix.

fn prefixes_of( &mut self, key: &E::Key ) -> Result<Self::ParentIterator<'_>, Self::Error>

Returns all entries that share a prefix with key, including the entry for key itself.

fn all(&mut self) -> Result<Self::RangeIterator<'_>, Self::Error>

Get all entries in the store

fn entry_remove(&mut self, key: &E::Key) -> Result<Option<E>, Self::Error>

Remove an entry from the store.

This will remove just the entry with the given key, but will not perform prefix deletion.

fn remove_prefix_filtered( &mut self, prefix: &E::Key, predicate: impl Fn(&E::Value) -> bool ) -> Result<usize, Self::Error>

Remove all entries whose key start with a prefix and for which the predicate callback returns true.

Returns the number of elements removed.

Provided Methods

fn get_range_len(&mut self, range: Range<E::Key>) -> Result<usize, Self::Error>

Returns the number of entries in the range.

Default impl is not optimized, but does avoid excessive memory usage.

fn initial_message(&mut self) -> Result<Message<E>, Self::Error>

Generates the initial message.

fn process_message<F, F2, F3>( &mut self, config: &SyncConfig, message: Message<E>, validate_cb: F, on_insert_cb: F2, content_status_cb: F3 ) -> Result<Option<Message<E>>, Self::Error>
where F: Fn(&Self, &E, ContentStatus) -> bool, F2: FnMut(&Self, E, ContentStatus), F3: Fn(&Self, &E) -> ContentStatus,

Processes an incoming message and produces a response. If terminated, returns None

validate_cb is called for each incoming entry received from the remote. It must return true if the entry is valid and should be stored, and false otherwise (which means the entry will be dropped and not stored).

on_insert_cb is called for each entry that was actually inserted into the store (so not for entries which validated, but are not inserted because they are older than one of their prefixes).

content_status_cb is called for each outgoing entry about to be sent to the remote. It must return a ContentStatus, which will be sent to the remote with the entry.

fn put(&mut self, entry: E) -> Result<InsertOutcome, Self::Error>

Insert a key value pair.

Entries are inserted if they compare strictly greater than all entries in the set of entries which have the same key as entry or have a key which is a prefix of entry.

Additionally, entries that have a key which is a prefix of the entry’s key and whose timestamp is not strictly greater than that of the new entry are deleted

Note: The deleted entries are simply dropped right now. We might want to make this return an iterator, to potentially log or expose the deleted entries.

Returns true if the entry was inserted. Returns false if it was not inserted.

Object Safety

This trait is not object safe.