iroh_docs::rpc
Module proto

Available on crate feature rpc only.

Protocol definitions for RPC.

Structs

AuthorCreateRequest

Create a new document author.

AuthorCreateResponse

Response for AuthorCreateRequest

AuthorDeleteRequest

Delete an author

AuthorDeleteResponse

Response for AuthorDeleteRequest

AuthorExportRequest

Exports an author

AuthorExportResponse

Response for AuthorExportRequest

AuthorGetDefaultRequest

Get the default author.

AuthorGetDefaultResponse

Response for AuthorGetDefaultRequest

AuthorImportRequest

Import author from secret key

AuthorImportResponse

Response to ImportRequest

AuthorListRequest

List document authors for which we have a secret key.

AuthorListResponse

Response for AuthorListRequest

AuthorSetDefaultRequest

Set the default author.

AuthorSetDefaultResponse

Response for AuthorSetDefaultRequest

CloseRequest

Open a document

CloseResponse

Response to CloseRequest

CreateRequest

Create a new document

CreateResponse

Response to CreateRequest

DelRequest

Delete entries in a document

DelResponse

Response to DelRequest

DocListRequest

List all documents

DocSubscribeRequest

Subscribe to events for a document.

DocSubscribeResponse

Response to DocSubscribeRequest

DropRequest

Stop the live sync for a doc, and optionally delete the document.

DropResponse

Response to DropRequest

ExportFileRequest

A request to the node to save the data of the entry to the given filepath

ExportFileResponse

Progress messages for an doc export operation

GetDownloadPolicyRequest

Get a download policy

GetDownloadPolicyResponse

Response to GetDownloadPolicyRequest

GetExactRequest

Get entries from a document

GetExactResponse

Response to GetExactRequest

GetManyRequest

Get entries from a document

GetManyResponse

Response to GetManyRequest

GetSyncPeersRequest

Get peers for document

GetSyncPeersResponse

Response to GetSyncPeersRequest

ImportFileRequest

A request to the node to add the data at the given filepath as an entry to the document

ImportFileResponse

Wrapper around ImportProgress.

ImportRequest

Import a document from a capability.

ImportResponse

Response to ImportRequest

LeaveRequest

Stop the live sync for a doc, and optionally delete the document.

LeaveResponse

Response to LeaveRequest

ListResponse

Response to DocListRequest

OpenRequest

Open a document

OpenResponse

Response to OpenRequest

RpcService

The RPC service type for the docs protocol.

SetDownloadPolicyRequest

Set a download policy

SetDownloadPolicyResponse

Response to SetDownloadPolicyRequest

SetHashRequest

Set an entry in a document via its hash

SetHashResponse

Response to SetHashRequest

SetRequest

Set an entry in a document

SetResponse

Response to SetRequest

ShareRequest

Share a document with peers over a ticket.

ShareResponse

The response to ShareRequest

StartSyncRequest

Start to sync a doc with peers.

StartSyncResponse

Response to StartSyncRequest

StatusRequest

Get info on a document

StatusResponse

Response to StatusRequest

Enums

Request

Response