iroh_docs
Module store

Storage trait and implementation for iroh-docs documents

Re-exports

pub use fs::Store;

Modules

fs

On disk storage for replicas.

Structs

FlatQuery

Query on all entries without aggregation.

MemPublicKeyStore

In-memory key storage

Query

Note: When using the SingleLatestPerKey query kind, the key filter is applied before the grouping, the author filter is applied after the grouping.

QueryBuilder

A query builder for document queries.

SingleLatestPerKeyQuery

Query that only returns the latest entry for a key which has entries from multiple authors.

Enums

AuthorFilter

Author matching.

DownloadPolicy

Download policy to decide which content blobs shall be downloaded.

FilterKind

Filter strategy used in download policies.

ImportNamespaceOutcome

Outcome of Store::import_namespace

KeyFilter

Key matching.

OpenError

Error return from Store::open_replica

SortBy

Fields by which the query can be sorted

SortDirection

Sort direction

Traits

DownloadPolicyStore

Store that gives read access to download policies for a document.

PublicKeyStore

Store trait for expanded public keys for authors and namespaces.