Trait iroh_docs::store::fs::ranges::RangeExt

pub trait RangeExt<K: Key, V: Value> {
    // Required methods
    fn next_map<T>(
        &mut self,
        map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> T
    ) -> Option<Result<T>>;
    fn next_filter_map<T>(
        &mut self,
        direction: &SortDirection,
        filter_map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> Option<T>
    ) -> Option<Result<T>>;

    // Provided method
    fn next_try_filter_map<T>(
        &mut self,
        direction: &SortDirection,
        filter_map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> Option<Result<T>>
    ) -> Option<Result<T>> { ... }
}

An extension trait for [Range] that provides methods for mapped retrieval.

Required Methods

fn next_map<T>( &mut self, map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> T ) -> Option<Result<T>>

Get the next entry and map the item with a callback function.

fn next_filter_map<T>( &mut self, direction: &SortDirection, filter_map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> Option<T> ) -> Option<Result<T>>

Get the next entry, but only if the callback function returns Some, otherwise continue.

With direction the range can be either process in forward or backward direction.

Provided Methods

fn next_try_filter_map<T>( &mut self, direction: &SortDirection, filter_map: impl for<'x> Fn(K::SelfType<'x>, V::SelfType<'x>) -> Option<Result<T>> ) -> Option<Result<T>>

Like Self::next_filter_map, but the callback returns a Result, and the result is flattened with the result from the range operation.

Object Safety

This trait is not object safe.