iroh_docs
Module sync

API for iroh-docs replicas

Structs

AuthorHeads

Timestamps of the latest entry for each author.

Entry

A single entry in a Replica

EntrySignature

Signature over an entry.

ReadOnly

Error that occurs trying to access the NamespaceSecret of a read-only Capability.

Record

The data part of an entry in a Replica.

RecordIdentifier

The identifier of a record.

Replica

Local representation of a mutable, synchronizable key-value store.

ReplicaInfo

In memory information about an open replica.

SignedEntry

A signed entry.

SyncOutcome

Outcome of a sync operation.

Enums

Capability

The capability of the namespace.

CapabilityError

Errors for capability operations

CapabilityKind

Kind of capability of the namespace.

ContentStatus

Whether the content status is available on a node.

Event

Event emitted by sync when entries are added.

InsertError

Error emitted when inserting entries into a Replica failed

InsertOrigin

Whether an entry was inserted locally or by a remote peer.

ValidationFailure

Reason why entry validation failed

Constants

MAX_TIMESTAMP_FUTURE_SHIFT

Max time in the future from our wall clock time that we accept entries for. Value is 10 minutes.

Type Aliases

ContentStatusCallback

Callback that may be set on a replica to determine the availability status for a content hash.

PeerIdBytes

Byte representation of a PeerId from iroh-net.

ProtocolMessage

Protocol message for the set reconciliation protocol.