Function iroh_docs::sync::validate_entry

fn validate_entry<S: Store<SignedEntry> + PublicKeyStore>(
    now: u64,
    store: &S,
    expected_namespace: NamespaceId,
    entry: &SignedEntry,
    origin: &InsertOrigin
) -> Result<(), ValidationFailure>

Validate a SignedEntry if it’s fit to be inserted.

This validates that

the entry’s author and namespace signatures are correct

the entry’s namespace matches the current replica

the entry’s timestamp is not more than 10 minutes in the future of our system time

the entry is newer than an existing entry for the same key and author, if such exists.